requirements.txt
# Core dependencies
kubernetes==28.1.0
numpy>=1.21.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Reinforcement Learning
stable-baselines3==2.2.1
gymnasium==0.29.1  # or gym==0.26.2 for older version
torch>=1.11.0
tensorboard>=2.10.0

# Utilities
pyyaml>=6.0
python-dateutil>=2.8.0
tqdm>=4.64.0

# Optional for advanced features
plotly>=5.10.0  # for interactive visualizations
requests>=2.28.0  # for HTTP operations